Crate tmi_rs

A WebSocket based, asynchronous Twitch chat client.

Example

#[macro_use]
extern crate log;

use std::env;
use std::error::Error;

use futures::future::{join, ready};
use futures::stream::StreamExt;
use futures::sink::SinkExt;

use tmi_rs::{TwitchChatConnection, TwitchClient, TwitchClientConfigBuilder};
use tmi_rs::client_messages::ClientMessage;
use tmi_rs::event::*;

#[tokio::main]
async fn main() -> Result<(), Box<dyn Error>> {
    env_logger::init();
    let channel = env::var("TWITCH_CHANNEL")?;
    let client: TwitchClient = TwitchClientConfigBuilder::default()
        .username(env::var("TWITCH_USERNAME")?)
        .token(env::var("TWITCH_AUTH")?)
        .build()? // build config
        .into(); // build client from config

    let TwitchChatConnection { mut sender, receiver, error_receiver } = client.connect().await?;

    // join a channel
    (&sender).send(ClientMessage::join(channel.clone())).await?;

    // process messages and do stuff with the data
    let process_messages = async {
        let send_result = receiver.filter_map(
            |event| {
                info!("{:?}", &event);
                ready(match &*event {
                    Event::PrivMsg(event_data) if event_data.message().starts_with("!hello") => {
                        // return response message to the stream
                        Some(ClientMessage::message(event_data.channel().to_owned(), "Hello World!"))
                    }
                    _ => None
                })
            })
            .map(Ok)
            .forward(&sender).await;

        if let Err(e) = send_result { error!("{}", e); }
    };

    // log any connection errors
    let process_errors = error_receiver.for_each(|error| async move {
        error!("Connection error: {}", error);
    });
    join(process_messages, process_errors).await;
    Ok(())
}

Re-exports

pub use client_messages::*;

Modules

client_messages
Types to represent messages sent by the client

event
Data types and parsing logic for events that can be received from the twitch servers.

irc
Parser for twitch flavored IRC

irc_constants
Constants containing the IRC reply codes

selectors
Convenience methods for matching specific event types. The functions in this module select an Arc<Event<T>> with a specific Event variant. The returned SelectorResult implements Deref for the inner type contained in the enum variant.

stream
Helpers and extension traits to deal with streams of received events and sent messages.

Structs

TwitchChatConnection
Contains the Streams and Sinks associated with an underlying websocket connection. They can be cloned freely to be shared across different tasks and threads.

TwitchChatSender
A wrapper around any Sink<ClientMessage<String>> that provides methods to send Twitch specific messages or commands to the sink.

TwitchClient
Represents a twitch chat client/connection. Call connect to establish a connection.

TwitchClientConfig
Holds the configuration for a twitch chat client. Convert it to a TwitchClient and call connect to establish a connection using it.

TwitchClientConfigBuilder
Builder for TwitchClientConfig.

Enums

Error
Error type for tmi-rs methods

Traits

StringRef
Trait that is used when generically referring to a &str, String, or other type that can be used like a borrowed string

Type Definitions

ChatReceiver
Receiver channel for all chat events

ChatSender
Sender to push messages to Twitch chat

ErrorReceiver
Receiver channel for connection and parsing errors