Enum tmi_rs::client_messages::ClientMessage

pub enum ClientMessage<T: StringRef> {
    PrivMsg {
        channel: T,
        message: T,
    },
    Whisper {
        recipient: T,
        message: T,
    },
    Join(T),
    Part(T),
    Nick(T),
    Pass(T),
    CapRequest(Vec<Capability>),
    Ping,
    Pong,
}

Messages to be sent from the client to twitch servers

Variants

PrivMsg

Fields of PrivMsg

channel: Tmessage: T

Whisper

Fields of Whisper

recipient: Tmessage: T

Join(T)

Part(T)

Nick(T)

Pass(T)

CapRequest(Vec<Capability>)

Ping

Pong

Methods

impl ClientMessage<String>

pub fn message<S1: Into<String> + Borrow<str>, S2: Into<String> + Borrow<str>>(
channel: S1,
message: S2
) -> Self

Send a channel message

pub fn join<S: Into<String> + Borrow<str>>(channel: S) -> Self

Joins a twitch channel

pub fn login<S: Into<String> + Borrow<str>>(
username: S,
token: S
) -> impl Stream<Item = Self>

Authenticates the user. Caution: this is normally called automatically when calling TwitchClient::connect, only use it if this stream was created in some other way.

pub fn ban<S: Into<String> + Borrow<str>>(channel: S, username: S) -> Self

Permanently ban a user

pub fn unban<S: Into<String> + Borrow<str>>(channel: S, username: S) -> Self

Unban a user

pub fn clear<S: Into<String> + Borrow<str>>(channel: S) -> Self

Clear a chat channel

pub fn color<S: Into<String> + Borrow<str>>(color: S) -> Self

Set the user name color

pub fn commercial<S: Into<String> + Borrow<str>>(
channel: S,
time: Option<usize>
) -> Self

Run a commercial

pub fn delete<S: Into<String> + Borrow<str>>(channel: S, msg_id: S) -> Self

Delete a specific message by its message ID tag.

pub fn disconnect<S: Into<String> + Borrow<str>>(&mut self) -> Self

Disconnect from chat

pub fn emote_only<S: Into<String> + Borrow<str>>(
channel: S,
status: bool
) -> Self

Set emote only mode on or off

pub fn followers_only<S: Into<String> + Borrow<str>>(
channel: S,
status: bool
) -> Self

Set followers only mode on or off

pub fn host<S: Into<String> + Borrow<str>>(channel: S, host_channel: S) -> Self

Host a channel

pub fn unhost<S: Into<String> + Borrow<str>>(channel: S) -> Self

Stop hosting

pub fn marker<S: Into<String> + Borrow<str>>(
channel: S,
description: Option<&str>
) -> Self

Set a stream marker

pub fn me<S: Into<String> + Borrow<str>>(message: S) -> Self

Send a /me message

pub fn make_mod<S: Into<String> + Borrow<str>>(channel: S, user: S) -> Self

Mod a user

pub fn unmod<S: Into<String> + Borrow<str>>(channel: S, user: S) -> Self

Unmod a user

pub fn r9k<S: Into<String> + Borrow<str>>(channel: S, on: bool) -> Self

Enable or disable r9k mode

pub fn raid<S: Into<String> + Borrow<str>>(channel: S, raid_channel: S) -> Self

Start a raid

pub fn unraid<S: Into<String> + Borrow<str>>(channel: S) -> Self

Stop a raid

pub fn slow<S: Into<String> + Borrow<str>>(channel: S, seconds: usize) -> Self

Enable slow mode with the given amount of seconds

pub fn slow_off<S: Into<String> + Borrow<str>>(channel: S) -> Self

Disable slow mode

pub fn subscribers<S: Into<String> + Borrow<str>>(channel: S, on: bool) -> Self

Enable or disable sub mode

pub fn timeout<S: Into<String> + Borrow<str>>(
channel: S,
user: S,
seconds: Option<usize>
) -> Self

Time a user out for the given amount of seconds or the default timeout if None is given

pub fn vip<S: Into<String> + Borrow<str>>(channel: S, user: S) -> Self

Make a user VIP in a channel

pub fn unvip<S: Into<String> + Borrow<str>>(channel: S, user: S) -> Self

Remove VIP status from a user

pub fn vips<S: Into<String> + Borrow<str>>(channel: S) -> Self

List the VIPs in a channel

pub fn whisper<S1: Into<String> + Borrow<str>, S2: Into<String> + Borrow<str>>(
user: S1,
message: S2
) -> Self

Send a whisper