Struct tmi_rs::event::EventData

pub struct EventData<T, Inner> where
    T: StringRef,
    Inner: Debug + Clone + Eq,  { /* fields omitted */ }

Content of a received message. Contains the sender, tags and and a generic Inner which contains the data specific to each event type.

Helper functions to access the contained data (message, target, specific tags) are implemented for for the EventContent<T, Inner> types they apply to.

Methods

impl<T, Inner> EventData<T, Inner> where
T: StringRef,
Inner: Debug + Clone + Eq,
Event<T>: From<EventData<T, Inner>>,

Methods common to all EventContent variants

pub fn sender(&self) -> &Option<T>

Get the sender of the message

pub fn event(&self) -> &Inner

Get the data of the inner event type