Module tmi_rs::event::tags

Convenience methods to access the IRCv3 tags twitch sends. It's possible to access the tag values using their names with the MessageTags<T> trait. However, there are accessor methods for all documented tags which are only implemented for the event types where they are expected to be set. Additionally, some of the methods parse the tags and return more easily processed structs instead of the raw string values (see EmoteTag::emotes()), EmoteSetsTag::emote_sets()).

Examples

use tmi_rs::event::{tags::UserDisplayTags, Event};

if let Event::PrivMsg(event) = parsed_message {
    println!("{:?}", event.color()); // Some("#0D4200")
}

Using the tag specific methods acts as a compile-time guard against trying to access tags on events where they are never defined:

use tmi_rs::event::Event;
use tmi_rs::event::tags::ClearChatTags;

if let Event::PrivMsg(event) = parsed_message {
    // only implemented for ClearChat events, won't compile
    println!("{:?}", event.ban_duration());
}

Structs

Badge
Badges from the badges and badges-info tags

EmoteReplacement
Replacement instruction for an emote

Traits

BadgeTags
Access to badges and badge-info tags

BitsTag
Access to bits tag

ClearChatTags
Tags specific to CLEARCHAT events

ClearMsgTags
Tags specific to CLEARMSG events

EmoteSetsTag
Access to the emote-sets tag

EmotesTag
Emote tag accessor

MessageTags
Access methods for individual IRC tags by their string tag keys

ModTag
Access to mod tag

RoomStateTags
Tags specific to ROOMSTATE events

UserDisplayTags
Access to color and display-name tags

UserIdTag
Access to the user-id tag

UserMessageTags
Tags that apply to both PRIVMSG and USERNOTICE.

UserNoticeTags
Tags specific to USERNOTICE messages

WhisperTags
Acessors for tags on whisper messages