Trait tmi_rs::event::tags::RoomStateTags

pub trait RoomStateTags<T: StringRef>: MessageTags<T> {
    fn emote_only(&self) -> bool { ... }

    fn followers_only(&self) -> Option<isize> { ... }

    fn r9k(&self) -> bool { ... }

    fn slow(&self) -> Option<usize> { ... }

    fn subs_only(&self) -> bool { ... }
}

Tags specific to ROOMSTATE events

Provided methods

fn emote_only(&self) -> bool

emote-only tag. Set when emote only mode is active.

fn followers_only(&self) -> Option<isize>

followers-only tag. Some(amount of minutes followed) when followers only mode is active, None when inactive.

fn r9k(&self) -> bool

r9k tag. Set when r9k mode is active.

fn slow(&self) -> Option<usize>

slow tag. Set to the number of seconds set for slow mode if active.

fn subs_only(&self) -> bool

subs-only tag. Set when emote only mode is active.

Loading content...