Struct tmi_rs::irc::IrcMessage

pub struct IrcMessage<T: StringRef> {
    pub tags: Option<FnvHashMap<T, T>>,
    pub prefix: Option<IrcPrefix<T>>,
    pub command: T,
    pub command_params: Vec<T>,
}

Represents an IRC message

Fields
tags: Option<FnvHashMap<T, T>>

IRCv3 tags

prefix: Option<IrcPrefix<T>>

IRC prefix

command: T

IRC Command name

command_params: Vec<T>

Command parameters

Methods

impl<'_> IrcMessage<&'_ str>

pub fn parse_many(input: &str) -> IResult<&str, Vec<IrcMessage<&str>>>

Parse one or more IRC messages into a Vec

pub fn parse(input: &str) -> IResult<&str, IrcMessage<&str>>

Parse a single IRC message

impl<T> IrcMessage<T> where
T: StringRef,

pub fn try_param(&self, index: usize) -> Result<&T, Error>

Get a reference to a command parameter by index, fails with an error if the parameter is not found

pub fn param(&self, index: usize) -> &T

Get a reference to a command parameter by index, panics if not set

pub fn params(&self) -> &[T]

Get a slice of all command parameters

pub fn sender(&self) -> Option<&T>

Get the name of the message sender from the IRC prefix, can be either the user or the nickname field depending on which one is set. For Twitch they should both be the same if both are given

pub fn host(&self) -> Option<&T>

Get the host part of the message, usually .tmi.twitch.tv