Trait tmi_rs::rate_limits::RateLimitable

pub trait RateLimitable {
    fn channel_limits(&self) -> Option<&str>;

    fn poll(&self, rate_limiter: &RateLimiter, cx: &mut Context) -> Poll<()> { ... }
}

Trait to apply to messages that contains information about which rate limits apply to the message

Required methods

fn channel_limits(&self) -> Option<&str>

Should return a channel name, if available. If a channel is returned, applies the slow mode and rate limit buckets configured for that channel

Loading content...
Provided methods

fn poll(&self, rate_limiter: &RateLimiter, cx: &mut Context) -> Poll<()>

Poll for sending the item using the given rate limiter instance

Loading content...