Trait tmi_rs::stream::SendStreamExt

pub trait SendStreamExt: Stream<Item = ClientMessage<String>> {
    fn rate_limited(
        self, 
        capacity: usize, 
        rate_limiter: &Arc<RateLimiter>
    ) -> BufferedRateLimiter<Self, ClientMessage<String>>
    where
        Self: Sized + Unpin,
    { ... }

    fn dedup(self) -> DedupMessages<Self>
    where
        Self: Sized + Unpin,
    { ... }

    fn split_oversize(self, max_len: usize) -> SplitOversize<Self>
    where
        Self: Sized + Unpin,
    { ... }
}

Extension trait with function to manipulate the stream of outgoing messages

Provided methods

fn rate_limited(
self,
capacity: usize,
rate_limiter: &Arc<RateLimiter>
) -> BufferedRateLimiter<Self, ClientMessage<String>> where
Self: Sized + Unpin,

Composes a fixed size buffered rate limiter in front of this sink. Messages passed into the sink are treated according to their implementations of RateLimitable. Messages that report rate limiting applies are put into a buffer which is sent

fn dedup(self) -> DedupMessages<Self> where
Self: Sized + Unpin,

Automatically add \u{0} to de-duplicate subsequent messages, to bypass Twitch's 30 second duplicate message detection

fn split_oversize(self, max_len: usize) -> SplitOversize<Self> where
Self: Sized + Unpin,

Splits messages over the given size limit into separate messages

Loading content...