Struct tmi_rs::TwitchChatSender

pub struct TwitchChatSender<Si: Clone + Unpin> { /* fields omitted */ }

A wrapper around any Sink<ClientMessage<String>> that provides methods to send Twitch specific messages or commands to the sink.

Methods

impl<Si> TwitchChatSender<Si> where
Si: Sink<ClientMessage<String>> + Unpin + Clone,

pub fn new(sink: Si) -> Self

Create a new chat sender using an underlying channel.

pub async fn message<'_, S1: Into<String> + Borrow<str>, S2: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S1,
message: S2
) -> Result<(), Error>

Send a whisper

pub async fn join<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S
) -> Result<(), Error>

Joins a twitch channel

pub async fn login<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
username: S,
token: S
) -> Result<(), Error>

Authenticates the user. Caution: this is normally called automatically when calling TwitchClient::connect, only use it if this stream was created in some other way.

pub async fn ban<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
username: S
) -> Result<(), Error>

Permanently ban a user

pub async fn unban<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
username: S
) -> Result<(), Error>

Unban a user

pub async fn clear<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S
) -> Result<(), Error>

Clear a chat channel

pub async fn color<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
color: S
) -> Result<(), Error>

Set the user name color

pub async fn commercial<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
time: Option<usize>
) -> Result<(), Error>

Run a commercial

pub async fn delete<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
msg_id: S
) -> Result<(), Error>

Delete a specific message by its message ID tag.

pub async fn disconnect<'_, S: Into<String> + Borrow<str>>(
&'_ mut self
) -> Result<(), Error>

Disconnect from chat

pub async fn emote_only<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
status: bool
) -> Result<(), Error>

Set emote only mode on or off

pub async fn followers_only<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
status: bool
) -> Result<(), Error>

Set followers only mode on or off

pub async fn host<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
host_channel: S
) -> Result<(), Error>

Host a channel

pub async fn unhost<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S
) -> Result<(), Error>

Stop hosting

pub async fn marker<'_, '_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
description: Option<&'_ str>
) -> Result<(), Error>

Set a stream marker

pub async fn me<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
message: S
) -> Result<(), Error>

Send a /me message

pub async fn make_mod<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
user: S
) -> Result<(), Error>

Mod a user

pub async fn unmod<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
user: S
) -> Result<(), Error>

Unmod a user

pub async fn r9k<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
on: bool
) -> Result<(), Error>

Enable or disable r9k mode

pub async fn raid<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
raid_channel: S
) -> Result<(), Error>

Start a raid

pub async fn unraid<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S
) -> Result<(), Error>

Stop a raid

pub async fn slow<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
seconds: usize
) -> Result<(), Error>

Enable slow mode with the given amount of seconds

pub async fn slow_off<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S
) -> Result<(), Error>

Disable slow mode

pub async fn subscribers<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
on: bool
) -> Result<(), Error>

Enable or disable sub mode

pub async fn timeout<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
user: S,
seconds: Option<usize>
) -> Result<(), Error>

Time a user out for the given amount of seconds or the default timeout if None is given

pub async fn vip<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
user: S
) -> Result<(), Error>

Make a user VIP in a channel

pub async fn unvip<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S,
user: S
) -> Result<(), Error>

Remove VIP status from a user

pub async fn vips<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
channel: S
) -> Result<(), Error>

List the VIPs in a channel

pub async fn whisper<'_, S: Into<String> + Borrow<str>>(
&'_ mut self,
user: S,
message: S
) -> Result<(), Error>

Send a whisper