Crate tmi_rs

A WebSocket based, asynchronous Twitch chat client.

Example

#![feature(async_closure)]
#[macro_use]
extern crate log;

use std::env;
use std::error::Error;
use tmi_rs::{TwitchClientBuilder, TwitchClient, futures_util::StreamExt};
use tmi_rs::event::{Event, ChannelMessageEventData};

#[tokio::main]
async fn main() -> Result<(), Box<dyn Error>> {
    env_logger::init();
    let channel = env::var("TWITCH_CHANNEL")?;
    let client: TwitchClient = TwitchClientBuilder::default()
        .username(env::var("TWITCH_USERNAME")?)
        .token(env::var("TWITCH_AUTH")?)
        .cap_membership(true)
        .build()?;
    let (mut sender, mut receiver) = client.connect().await?;

    sender.join(channel.clone()).await?;

    while let Some(event) = receiver.next().await {
        match &*event {
            Ok(event) => {
                match event {
                    Event::PrivMsg(event_data) => {
                        if event_data.message().starts_with("!hello") {
                            sender.message(event_data.channel().to_owned(), "Hello World!").await?;
                        }
                    }
                    _ => {
                        info!("Event received: {:?}", event)
                    }
                }
            }
            Err(e) => error!("Connection error: {}", e)
        }
    }
    receiver.for_each(async move |event| {
        info!("{:?}", event);
    }).await;
    Ok(())
}

Re-exports

pub use futures_channel;

pub use futures_core;

pub use futures_sink;

pub use futures_util;

Modules

client_messages
Types to represent messages sent by the client

event
Data types and parsing logic for events that can be received from the twitch servers.

irc
Parser for twitch flavored IRC

irc_constants
Constants containing the IRC reply codes

selectors
Convenience functions for matching specific event types

Structs

TwitchChatSender
A wrapper around any Sink<ClientMessage<Cow<'static, str>>> that provides methods to send Twitch specific messages or commands to the sink.

TwitchClient
Holds the configuration for a twitch chat client. Call connect to establish a connection using it.

TwitchClientBuilder
Builder for TwitchClient.

Enums

Error
Error type for tmi-rs methods

Traits

StringRef
Trait that is used when generically referring to a &str, String, or other type that can be used like a borrowed string